upd
added no-scroll class on body on toggle of modal-window
functions located in separated js-files
