+++ src/index.js
@@ -1,3 +1,5 @@
 import './sass/main.scss';
 import './js/api-service';
 import './js/modal-window';
+import './js/render-movies';
+import './js/make-markup';
