+++ src/sass/components/_header.scss
@@ -38,7 +38,7 @@
 .header-top {
   display: flex;
   justify-content: space-between;
-  align-items: flex-end;
+  align-items: center;
   margin-bottom: 40px;
 }
